requirements.txt
# Telegram Bot
python-telegram-bot==20.3
telethon==1.31.1

# Payment Gateway
requests==2.31.0
aiohttp==3.8.5

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Environment & Config
python-dotenv==1.0.0
pydantic==2.4.2

# Utilities
aiofiles==23.2.1
Pillow==10.0.1
qrcode==7.4.2

# Async & Concurrency
aioredis==2.0.1
celery==5.3.4

# Logging
python-json-logger==2.0.7

# Security
cryptography==41.0.7
PyJWT==2.8.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.12.0
flake8==6.1.0
